登录消息定义

+++ src/main/java/cn/loach/server/serializable/LoachSerializable.java
@@ -4,10 +4,10 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class LoachSerializable {
-    public static final int JAVA_SERIALIZABLE_TYPE = 1;
-    public static final int JSON_SERIALIZABLE_TYPE = 2;
+    public static final byte JAVA_SERIALIZABLE_TYPE = 1;
+    public static final byte JSON_SERIALIZABLE_TYPE = 2;
 
-    private static final Map<Integer, SerializableAlgorithmEnum> serializableTypeMap = new HashMap<>();
+    private static final Map<Byte, SerializableAlgorithmEnum> serializableTypeMap = new HashMap<>();
 
     static {
         serializableTypeMap.put(JAVA_SERIALIZABLE_TYPE, SerializableAlgorithmEnum.JAVA_SERIALIZABLE);
@@ -15,7 +15,7 @@ public class LoachSerializable {
     }
 
 
-    public static SerializableAlgorithmEnum getSerializable(int serializableType) {
+    public static SerializableAlgorithmEnum getSerializable(byte serializableType) {
         return serializableTypeMap.get(serializableType);
     }
 }
